requirements.txt
aiogram==3.15.0
aiohttp==3.11.11
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic==2.10.4
uvicorn==0.34.0
fastapi==0.115.6
